requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-ninja==1.1.0
djangorestframework==3.14.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
idna==3.6
pyasn1==0.5.1
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
